ai_service/requirements.txt
typing_extensions
typing
langchain_openai
langchain_core
langchain
langgraph
python-dotenv
fastapi
pydantic
uvicorn